pyproject.toml
[tool.poetry]
name = "latest.cat"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.14.1"
uvicorn = {extras = ["standard"], version = "^0.13.3"}
databases = "^0.4.1"
aiosqlite = "^0.16.0"
pydantic = "^1.7.3"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
flake8 = "^3.8.4"
PyGithub = "^1.54.1"
PyYAML = "^5.4.1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true
